Updated CRUD operations

+++ src/Models/Model.php
@@ -6,7 +6,7 @@ use Doctrine\DBAL\Connection;
 use Doctrine\DBAL\Exception;
 use Doctrine\DBAL\Query\QueryBuilder;
 use Nexa\Attributes\Common\PrimaryKey;
-use Nexa\Attributes\Entities\Entity;
+use Nexa\Collections\Collection;
 use Nexa\Exceptions\NotFoundException;
 use Nexa\Nexa;
 use Nexa\Reflection\EntityReflection;
@@ -17,11 +17,31 @@ class Model
 
     private static Connection $connection;
 
+    private static Model $model;
+
     private static string $table;
 
     private static QueryBuilder $queryBuilder;
 
     private static ?string $primaryKey = null;
+
+    protected $entity;
+
+    protected $hidden;
+
+    protected $fillable;
+
+    protected $timestamp = false;
+
+    protected $soft_delete = false;
+
+    protected $date_format = "Y-m-d h:i:s";
+
+    protected $created_at = 'created_at';
+
+    protected $updated_at = 'updated_at';
+
+    protected $deleted_at = 'deleted_at';
 
 
     /**
@@ -30,36 +50,39 @@ class Model
     public function __construct()
     {
 
-        $reflection = new EntityReflection($this::class);
         self::$connection = Nexa::getConnection();
+        $reflection = new EntityReflection($this->entity); // TODO: create search_entity method
         self::$table = $reflection->getTable(Nexa::$inflector);
         self::$queryBuilder = self::$connection->createQueryBuilder();
         self::$primaryKey = self::getPrimaryKey($reflection);
+        self::$model = $this;
     }
 
     /**
      * @throws Exception
      */
-    public static function find($id, $columns = ["*"]): array|false
+    public static function find($id): Collection | false
     {
         new static;
 
-        return self::$queryBuilder->select(implode(",", $columns))
+        $result =  self::$queryBuilder->select('*')
             ->from(self::$table)
             ->where(self::$primaryKey . "= ?")
             ->setParameters([$id])
             ->fetchAssociative();
+
+        return is_array($result) ? new Collection($result) : $result;
     }
 
     /**
      * @throws NotFoundException
      * @throws Exception
      */
-    public static function findOrFail($id, $columns = ["*"]): array
+    public static function findOrFail($id): Collection
     {
         new static;
 
-        $result = self::$queryBuilder->select(implode(",", $columns))
+        $result = self::$queryBuilder->select()
             ->from(self::$table)
             ->where(self::$primaryKey . "= ?")
             ->setParameters([$id])
@@ -70,19 +93,23 @@ class Model
             throw new NotFoundException('Resource not Found', 4004);
         }
 
-        return $result;
+        return new Collection($result);
     }
 
     /**
      * @throws Exception
      */
-    public static function findAll(array $columns = ["*"]): array
+    public static function findAll($columns = ["*"]): Collection
     {
         new static;
 
-        return self::$queryBuilder->select(implode(",", $columns))
+        $result = self::$queryBuilder->select(
+            is_array($columns) ? implode(',', $columns) : implode(',', func_get_args())
+        )
             ->from(self::$table)
             ->fetchAllAssociative();
+
+        return new Collection($result);
     }
 
     public static function like(string $column, string $search, $columns = ['*']): array|false
@@ -91,7 +118,7 @@ class Model
 
         return self::$queryBuilder->select(implode(",", $columns))
             ->from(self::$table)
-            ->where("$column LIKE '%$search%'")
+            ->where("$column LIKE '$search'")
             ->fetchAllAssociative();
     }
 
@@ -104,8 +131,21 @@ class Model
 
         $data = self::secure($data);
 
-        return self::$connection->insert(self::$table, $data);
+        if (implode('', self::$model->fillable) != "*") {
 
+            foreach ($data as $key => $value) {
+
+                if (!in_array($key, self::$model->fillable)) {
+                    unset($data[$key]);
+                }
+            }
+        }
+        
+        if (self::$model->timestamp) {
+            $data[self::$model->created_at] = date(self::$model->date_format);
+        }
+
+        return self::$connection->insert(self::$table, $data);
     }
 
     /**
@@ -117,8 +157,11 @@ class Model
 
         $data = self::secure($data);
 
-        return self::$connection->update(self::$table, $data, $conditions);
+        if (self::$model->timestamp) {
+            $data[self::$model->updated_at] = date(self::$model->date_format);
+        }
 
+        return self::$connection->update(self::$table, $data, $conditions);
     }
 
     /**
@@ -130,8 +173,18 @@ class Model
 
         $id = self::secure($id);
 
-        return self::$connection->delete(self::$table, [self::$primaryKey => $id]);
+        if (self::$model->soft_delete) {
 
+            return self::$connection->update(
+                self::$table,
+                [
+                    self::$model->deleted_at => date(self::$model->date_format)
+                ],
+                [self::$primaryKey => $id]
+            );
+        }
+
+        return self::$connection->delete(self::$table, [self::$primaryKey => $id]);
     }
 
     /**
@@ -141,19 +194,18 @@ class Model
     {
         new static;
 
-        return self::$connection->delete(self::$table, $conditions);
-
+        return self::$connection->delete(self::$table, self::secure($conditions));
     }
 
-    private static function secure(array | string $data) {
+    private static function secure(array | string $data)
+    {
 
         if (is_array($data)) {
 
-            return array_map(function($value) {
+            return array_map(function ($value) {
 
-               return self::secure($value);
+                return self::secure($value);
             }, $data);
-
         }
         return htmlspecialchars($data, ENT_QUOTES, 'UTF-8');
     }
@@ -163,10 +215,9 @@ class Model
 
         $properties = $reflection->getProperties();
 
-        $result = array_filter($properties, function($property) {
+        $result = array_filter($properties, function ($property) {
             return $property->getAttributes(PrimaryKey::class);
         });
-        return count($result) > 0 ? $result[0]->getName() : null;
+        return count($result) > 0 ? $result[0]->getName() : 'id';
     }
-
 }
